requirements.txt
streamlit
pytesseract
python-docx
pymupdf
Pillow
google-generativeai
